Refactor kwargs functionality with unified tests and helpers
