--- tests/examples/test_functions_with_kwargs.http.json
@@ -1,74 +1,0 @@
-{
-    "fixtures": [
-        "dict_value"
-    ],
-    "stages": [
-        {
-            "name": "test_functions_with_kwargs_basic",
-            "url": "https://httpbin.org/json",
-            "save": {
-                "vars": {
-                    "test_var": "json.slideshow.title"
-                },
-                "functions": [
-                    {
-                        "function": "test_helpers:extract_custom_data",
-                        "kwargs": {
-                            "field_path": "slideshow.author",
-                            "default_value": "unknown_author"
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_functions_with_kwargs_multiple_fields",
-            "url": "https://httpbin.org/json",
-            "save": {
-                "functions": [
-                    {
-                        "function": "test_helpers:extract_multiple_fields",
-                        "kwargs": {
-                            "fields": ["slideshow.title", "slideshow.author", "slideshow.date"]
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_functions_with_kwargs_filter",
-            "url": "https://httpbin.org/json",
-            "save": {
-                "functions": [
-                    {
-                        "function": "test_helpers:extract_with_filter",
-                        "kwargs": {
-                            "field_path": "slideshow.slides",
-                            "filter_field": "type",
-                            "filter_value": "all"
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_functions_mixed_format",
-            "url": "https://httpbin.org/json",
-            "save": {
-                "vars": {
-                    "test_var": "json.slideshow.title"
-                },
-                "functions": [
-                    "test_helpers:extract_test_data",
-                    {
-                        "function": "test_helpers:extract_custom_data",
-                        "kwargs": {
-                            "field_path": "slideshow.author",
-                            "default_value": "custom_author"
-                        }
-                    }
-                ]
-            }
-        }
-    ]
-}
