--- tests/examples/test_verify_functions_with_kwargs.http.json
@@ -1,81 +1,0 @@
-{
-    "stages": [
-        {
-            "name": "test_verify_functions_with_kwargs_basic",
-            "url": "https://httpbin.org/json",
-            "verify": {
-                "status": 200,
-                "functions": [
-                    {
-                        "function": "test_verify_helpers:verify_response_status_custom",
-                        "kwargs": {
-                            "expected_status": 200
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_verify_functions_with_kwargs_text",
-            "url": "https://httpbin.org/json",
-            "verify": {
-                "functions": [
-                    {
-                        "function": "test_verify_helpers:verify_response_contains_text",
-                        "kwargs": {
-                            "expected_text": "Sample Slide Show",
-                            "case_sensitive": false
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_verify_functions_with_kwargs_header",
-            "url": "https://httpbin.org/headers",
-            "verify": {
-                "functions": [
-                    {
-                        "function": "test_verify_helpers:verify_response_header_value",
-                        "kwargs": {
-                            "header_name": "content-type",
-                            "expected_value": "application/json"
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_verify_functions_with_kwargs_json",
-            "url": "https://httpbin.org/json",
-            "verify": {
-                "functions": [
-                    {
-                        "function": "test_verify_helpers:verify_response_json_field",
-                        "kwargs": {
-                            "field_path": "slideshow.title",
-                            "expected_value": "Sample Slide Show"
-                        }
-                    }
-                ]
-            }
-        },
-        {
-            "name": "test_verify_functions_mixed_format",
-            "url": "https://httpbin.org/json",
-            "verify": {
-                "status": 200,
-                "functions": [
-                    "test_verify_helpers:verify_response_has_json",
-                    {
-                        "function": "test_verify_helpers:verify_response_json_field",
-                        "kwargs": {
-                            "field_path": "slideshow.author",
-                            "expected_value": "Yours Truly"
-                        }
-                    }
-                ]
-            }
-        }
-    ]
-}
